❇️ better management setup for route handlers
